test/requirements/requirements.txt
# Dependencias a serem instaladas no projeto
nose
behave
selenium
behave-html-formatter
allure-behave
faker
Import-Module